[test] simple pplication test  code

+++ CppLibrary/include/aiRunner.h
@@ -9,5 +9,5 @@ extern "C" bool InitModel(const char* modelPath, bool cpu_use);
 extern "C" void RunModelInt(int* data, size_t num_elements);
 extern "C" void RunModelFloat(float* data, size_t num_elements);
 extern "C" void RunModelDouble(double* data, size_t num_elements);
-extern "C" const float* GetFlattenedOutput(size_t* size);
-extern "C" const int64_t* GetOriginalShape(size_t* size);
+extern "C" const float* GetFlattenedOutput(int* size);
+extern "C" const int64_t* GetOriginalShape(int* size);
